Feature/ground labels (#35)
* add optional extra categories before objects

* swap packed order of instance and category ids

* add extra metadata config and publish correct invalid list

* make alpha value uint8_t

+++ semantic_inference_ros/include/semantic_inference_ros/pointcloud_projection.h
@@ -30,7 +30,7 @@ struct ProjectionConfig {
   //! Input label remapping for input pointcloud
   std::unordered_map<uint32_t, uint32_t> input_remapping;
   //! Alpha value for out-of-view point colors
-  uint16_t out_of_view_alpha = 0;
+  uint8_t out_of_view_alpha = 0;
   //! Whether the input labels are in the form of instance IDs (16 bit semantics + 16
   //! bit instance)
   bool instance_id = false;
